Управління камерою, створення камери винесено у окремий клас
